requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
jinja2==3.1.4
reportlab==4.2.2
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1